test_to_period
